+++ hermes-integration-test-app/build.gradle
@@ -11,7 +11,7 @@ buildscript {
 }
 
 version "0.1"
-group "hermes.integration.test.app"
+group "com.binxhealth"
 
 apply plugin:"eclipse"
 apply plugin:"idea"
