+++ hermes-integration-test-app/grails-app/controllers/com/binxhealth/hermes/test/app/UrlMappings.groovy
@@ -1,0 +1,16 @@
+package com.binxhealth.hermes.test.app
+
+class UrlMappings {
+
+    static mappings = {
+        "/$controller/$action?/$id?(.$format)?"{
+            constraints {
+                // apply constraints here
+            }
+        }
+
+        "/"(view:"/index")
+        "500"(view:'/error')
+        "404"(view:'/notFound')
+    }
+}
