--- hermes-integration-test-app/grails-app/controllers/hermes/integration/test/app/UrlMappings.groovy
@@ -1,16 +1,0 @@
-package hermes.integration.test.app
-
-class UrlMappings {
-
-    static mappings = {
-        "/$controller/$action?/$id?(.$format)?"{
-            constraints {
-                // apply constraints here
-            }
-        }
-
-        "/"(view:"/index")
-        "500"(view:'/error')
-        "404"(view:'/notFound')
-    }
-}
