+++ hermes-integration-test-app/grails-app/init/com/binxhealth/hermes/test/app/Application.groovy
@@ -1,0 +1,10 @@
+package com.binxhealth.hermes.test.app
+
+import grails.boot.GrailsApp
+import grails.boot.config.GrailsAutoConfiguration
+
+class Application extends GrailsAutoConfiguration {
+    static void main(String[] args) {
+        GrailsApp.run(Application, args)
+    }
+}
