+++ hermes-integration-test-app/grails-app/init/com/binxhealth/hermes/test/app/BootStrap.groovy
@@ -1,0 +1,9 @@
+package com.binxhealth.hermes.test.app
+
+class BootStrap {
+
+    def init = { servletContext ->
+    }
+    def destroy = {
+    }
+}
