--- hermes-integration-test-app/grails-app/init/hermes/integration/test/app/Application.groovy
@@ -1,10 +1,0 @@
-package hermes.integration.test.app
-
-import grails.boot.GrailsApp
-import grails.boot.config.GrailsAutoConfiguration
-
-class Application extends GrailsAutoConfiguration {
-    static void main(String[] args) {
-        GrailsApp.run(Application, args)
-    }
-}
