--- hermes-integration-test-app/grails-app/init/hermes/integration/test/app/BootStrap.groovy
@@ -1,9 +1,0 @@
-package hermes.integration.test.app
-
-class BootStrap {
-
-    def init = { servletContext ->
-    }
-    def destroy = {
-    }
-}
