+++ hermes-plugin/build.gradle
@@ -11,7 +11,7 @@ buildscript {
 }
 
 version "1.3.1"
-group "com.atlasgenetics"
+group "com.binxhealth"
 
 apply plugin:"eclipse"
 apply plugin:"idea"
